lambdas/auth/authorize/requirements.txt
boto3==1.22.6
requests==2.27.1
pyjwt==2.4.0
cryptography==37.0.2